chore:1차 리뷰 수정사항 반영(#149)

+++ src/main/java/com/kakaotech/team18/backend_server/domain/notices/service/NoticeService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface NoticeService {
 
-    List<NoticeBriefResponseDto> getAllNotices(Long page, Long size);
+    List<NoticeBriefResponseDto> getAllNotices(Integer page, Integer size);
 
     NoticeResponseDto getNoticeById(Long noticeId);
